% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsAndStuff.R
\name{clogitRV}
\alias{clogitRV}
\title{Conditional Logistic Regression with Robust Variance}
\usage{
clogitRV(formula, id.set, data)
}
\arguments{
\item{formula}{Similar to a formula for glm.}

\item{id.set}{The ID variable for matched sets.}

\item{data}{The dataset}
}
\value{
coef   The coefficeints (i.e. betas) and their standard error

cov    The covariance matrix of the coefficeints (useful for wald tests)
}
\description{
Performs conditional logistic regression with a robust variance estimator
}
\examples{
n    <-  10
Y    <- rep(c(1,0,1,0),n)
ID   <- rep(c(1:n),each=4)
X1   <- rnorm(4*n)
X2   <- rnorm(4*n)
data  <- cbind(Y,X1,X2,ID)
clogitRV(Y~X1+X2,~ID,data)
}
