% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsAndStuff.R
\docType{data}
\name{simData}
\alias{simData}
\title{A Simulated Dataset}
\format{
A data frame with 100 matched case/control pairs with two measurements per person
\describe{
\item{Y}{Case/Control Status}
\item{x}{biomarker}
\item{t}{time between biomarker measurement and case diagnosis}
\item{z}{a covariate}
}
}
\usage{
simData
}
\description{
A simulated dataset for showing a time-varying effect
}
\keyword{datasets}
