% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsAndStuff.R
\name{clogitRV}
\alias{clogitRV}
\title{Add together two numbers.}
\usage{
clogitRV(formula, id.set, data)
}
\arguments{
\item{formula}{Similar to a formula for logistic regression.}

\item{id.set}{The ID variable for matched sets.}
}
\value{
The coefficeints, standard errors, and the covariance matrix for the estimates.
}
\description{
This is a description
}
\examples{
n    <-  10
Y    <- rep(c(1,0,1,0),n)
ID   <- rep(c(1:n),each=4)
X1   <- rnorm(4*n)
X2   <- rnorm(4*n)
data  <- cbind(Y,X1,X2,ID)
clogitRV(Y~X1+X2,~ID,data)
}
